requirements.txt
selenium==4.11.2
pytest==7.4.0
pytest-html==3.2.0
openpyxl==3.1.2
pyyaml==6.0